pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm[toml]>=8.0.4"]
build-backend = "setuptools.build_meta"

[project]
name = "niftyone"
description = "Large-scale neuroimaging visualization using FiftyOne"
authors = [{name = "Connor Lane", email = "connor.lane858@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8, <3.13"
license = {text = "MIT License"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "av",
  "bids2table",
  "elbow",
  "fiftyone",
  "matplotlib",
  "nibabel",
  "nilearn",
  "numpy",
  "pandas >= 2.0, < 2.1",
  "Pillow",
  "scikit-learn",
  "scipy"
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["pytest>=8.0.2", "pytest-cov>=4.1.0"]
dev = [
  "mypy>=1.9.0",
  "pre-commit>=3.5.0",
  "ruff>=0.3.7",
  "setuptools-scm>=8.0.4"
]

[project.urls]
"Homepage" = "https://github.com/childmindresearch/niftyone"

[project.scripts]
niftyone = "niftyone.__main__:main"

[tool.setuptools_scm]
write_to = "src/niftyone/_version.py"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"

[tool.mypy]
ignore_missing_imports = true
allow_redefinition = true
disable_error_code = ["no-redef"]

[tool.ruff]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py38"

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
  "ANN101",  # self should not be annotated.
  "ANN102",  # cls should not be annotated.
  "D105",  # undocumented-magic-method (e.g. __str__)
  "D107"  # undocumented-public-init (e.g. __init__)
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
